+++ Anviz.SDK/Anviz.SDK.csproj
@@ -51,7 +51,7 @@
     <Compile Include="Commands\GetTCPParametersCommand.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Record.cs" />
-    <Compile Include="Response.cs" />
+    <Compile Include="Responses\Response.cs" />
     <Compile Include="Statistic.cs" />
     <Compile Include="TcpParameters.cs" />
     <Compile Include="UserInfo.cs" />
