+++ Anviz.SDK/AnvizManager.cs
@@ -1,4 +1,5 @@
 ﻿using Anviz.SDK.Commands;
+using Anviz.SDK.Responses;
 using Anviz.SDK.Utils;
 using System;
 using System.Collections.Generic;
